refactor: VO -> DTO로 이름 변경, record 활용
