+++ src/main/java/bridge/common/dto/GameResult.java
@@ -1,0 +1,7 @@
+package bridge.common.dto;
+
+public record GameResult(
+		MoveResult moveResult,
+		int totalTryCount
+) {
+}
